version update

+++ src/main/java/de/j4velin/systemappmover/AppClickListener.java
@@ -16,14 +16,9 @@
 
 package de.j4velin.systemappmover;
 
-import java.io.File;
-import java.util.List;
-
-import com.stericson.RootTools.RootTools;
-
 import android.app.ActivityManager;
-import android.app.AlertDialog;
 import android.app.ActivityManager.RunningAppProcessInfo;
+import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.DialogInterface.OnClickListener;
@@ -35,6 +30,11 @@ import android.view.View;
 import android.widget.AdapterView;
 import android.widget.AdapterView.OnItemClickListener;
 
+import com.stericson.RootTools.RootTools;
+
+import java.io.File;
+import java.util.List;
+
 public class AppClickListener implements OnItemClickListener {
 
     private final AppPicker ap;
@@ -44,8 +44,7 @@ public class AppClickListener implements OnItemClickListener {
     }
 
     public void onItemClick(final AdapterView<?> parent, final View view, final int position, long id) {
-        if (position >= ap.apps.size())
-            return;
+        if (position >= ap.apps.size()) return;
 
         if ("MOVED".equals(view.getTag())) {
             ap.activity.showErrorDialog("Please reboot before moving this app again");
@@ -56,14 +55,13 @@ public class AppClickListener implements OnItemClickListener {
         boolean tmpAlreadySys = (tmp.flags & ApplicationInfo.FLAG_SYSTEM) == 1;
 
         // update necessary?
-        if ((tmpAlreadySys && tmp.sourceDir.contains("/data/app/"))
-                || (!tmpAlreadySys && tmp.sourceDir.contains(MoverActivity.SYSTEM_APP_FOLDER))) {
+        if ((tmpAlreadySys && tmp.sourceDir.contains("/data/app/")) ||
+                (!tmpAlreadySys && tmp.sourceDir.contains(MoverActivity.SYSTEM_APP_FOLDER))) {
             try {
                 tmp = ap.pm.getApplicationInfo(tmp.packageName, 0);
             } catch (NameNotFoundException e1) {
                 ap.activity.showErrorDialog("App not found");
-                if (BuildConfig.DEBUG)
-                    Logger.log(e1);
+                if (BuildConfig.DEBUG) Logger.log(e1);
                 return;
             }
         }
@@ -72,26 +70,24 @@ public class AppClickListener implements OnItemClickListener {
         final String appName = (String) app.loadLabel(ap.pm);
         final boolean alreadySys = (app.flags & ApplicationInfo.FLAG_SYSTEM) == 1;
 
-        if (BuildConfig.DEBUG)
-            Logger.log("Trying to move " + appName + " - " + app.packageName);
+        if (BuildConfig.DEBUG) Logger.log("Trying to move " + appName + " - " + app.packageName);
 
         if (app.packageName.equals(ap.activity.getPackageName())) {
             ap.activity.showErrorDialog("Can not move myself");
-            if (BuildConfig.DEBUG)
-                Logger.log("Can not move myself");
+            if (BuildConfig.DEBUG) Logger.log("Can not move myself");
             return;
         }
 
         if (alreadySys && app.sourceDir.contains("/data/app/")) {
-            if (BuildConfig.DEBUG)
-                Logger.log("Need to remove updates first");
+            if (BuildConfig.DEBUG) Logger.log("Need to remove updates first");
             AlertDialog.Builder builder = new AlertDialog.Builder(ap.activity);
-            builder.setTitle("Error").setMessage("Can not move " + appName + ": Remove installed updates first.")
+            builder.setTitle("Error")
+                    .setMessage("Can not move " + appName + ": Remove installed updates first.")
                     .setPositiveButton("Remove updates", new DialogInterface.OnClickListener() {
                         public void onClick(final DialogInterface dialog, int id) {
                             try {
-                                ap.activity.startActivity(new Intent(Intent.ACTION_DELETE, Uri
-                                        .parse("package:" + app.packageName)));
+                                ap.activity.startActivity(new Intent(Intent.ACTION_DELETE,
+                                        Uri.parse("package:" + app.packageName)));
                                 dialog.dismiss();
                             } catch (Exception e) {
                             }
@@ -107,24 +103,22 @@ public class AppClickListener implements OnItemClickListener {
             builder.create().show();
             return;
         } else if (!alreadySys && tmp.sourceDir.contains(MoverActivity.SYSTEM_APP_FOLDER)) {
-            ap.activity.showErrorDialog("Can not move " + appName + ": Undefined app status. You might need to reboot once.");
-            if (BuildConfig.DEBUG)
-                Logger.log("Undefined app status: IsSystem = " + alreadySys + " path = " + tmp.sourceDir);
+            ap.activity.showErrorDialog("Can not move " + appName +
+                    ": Undefined app status. You might need to reboot once.");
+            if (BuildConfig.DEBUG) Logger.log(
+                    "Undefined app status: IsSystem = " + alreadySys + " path = " + tmp.sourceDir);
             return;
         }
 
         if (!alreadySys && app.sourceDir.endsWith("pkg.apk")) {
             if (app.sourceDir.contains("asec")) {
-                if (BuildConfig.DEBUG)
-                    Logger.log("Paid app? Path = " + app.sourceDir);
-                ap.activity
-                        .showErrorDialog(appName
-                                + " seems to be a paid app and therefore can not be converted to system app due to limitations by the Android system");
+                if (BuildConfig.DEBUG) Logger.log("Paid app? Path = " + app.sourceDir);
+                ap.activity.showErrorDialog(appName +
+                        " seems to be a paid app and therefore can not be converted to system app due to limitations by the Android system");
             } else {
-                if (BuildConfig.DEBUG)
-                    Logger.log("SD card? " + app.sourceDir);
-                ap.activity.showErrorDialog(appName
-                        + " is currently installed on SD card. Please move to internal memory before moving to /system/app/");
+                if (BuildConfig.DEBUG) Logger.log("SD card? " + app.sourceDir);
+                ap.activity.showErrorDialog(appName +
+                        " is currently installed on SD card. Please move to internal memory before moving to /system/app/");
             }
             return;
         }
@@ -138,36 +132,34 @@ public class AppClickListener implements OnItemClickListener {
                 if (RootTools.remount("/system", "rw")) {
                     try {
 
-                        if (BuildConfig.DEBUG)
-                            Logger.log("process name: " + app.processName);
+                        if (BuildConfig.DEBUG) Logger.log("process name: " + app.processName);
 
                         ActivityManager activityManager = (ActivityManager) ap.activity
                                 .getSystemService(Context.ACTIVITY_SERVICE);
-                        List<RunningAppProcessInfo> runningProcInfo = activityManager.getRunningAppProcesses();
+                        List<RunningAppProcessInfo> runningProcInfo =
+                                activityManager.getRunningAppProcesses();
                         String[] pkgList;
                         for (RunningAppProcessInfo p : runningProcInfo) {
                             pkgList = p.pkgList;
                             for (String pkg : pkgList) {
                                 if (pkg.equals(app.processName)) {
-                                    if (BuildConfig.DEBUG)
-                                        Logger.log("killing: " + p.processName);
+                                    if (BuildConfig.DEBUG) Logger.log("killing: " + p.processName);
                                     RootTools.killProcess(p.processName);
                                     break;
                                 }
                             }
                         }
 
-                        if (BuildConfig.DEBUG)
-                            Logger.log("source: " + app.sourceDir);
+                        if (BuildConfig.DEBUG) Logger.log("source: " + app.sourceDir);
 
                         String mvcmd, newFile;
-                        List<String> output = null;
+                        List<String> output;
                         if (!alreadySys) {
                             if (app.sourceDir.endsWith("/pkg.apk")) {
-                                newFile = MoverActivity.SYSTEM_APP_FOLDER + app.packageName + "-asec.apk";
+                                newFile = MoverActivity.SYSTEM_APP_FOLDER + app.packageName +
+                                        "-asec.apk";
                                 if (!RootTools.remount("/mnt", "rw")) {
-                                    if (BuildConfig.DEBUG)
-                                        Logger.log("Can not remount /mnt");
+                                    if (BuildConfig.DEBUG) Logger.log("Can not remount /mnt");
                                     ap.activity.showErrorDialog("Can not remount /mnt/asec");
                                     return;
                                 }
@@ -175,20 +167,22 @@ public class AppClickListener implements OnItemClickListener {
                                 if (BuildConfig.DEBUG)
                                     Logger.log("source ends with /pkg.apk -> paid app");
                             } else {
-                                newFile = app.sourceDir.replace("/data/app/", MoverActivity.SYSTEM_APP_FOLDER);
-                                mvcmd = "busybox mv " + app.sourceDir + " " + MoverActivity.SYSTEM_APP_FOLDER;
+                                newFile = app.sourceDir
+                                        .replace("/data/app/", MoverActivity.SYSTEM_APP_FOLDER);
+                                mvcmd = "busybox mv " + app.sourceDir + " " +
+                                        MoverActivity.SYSTEM_APP_FOLDER;
                             }
                         } else {
                             if (app.sourceDir.endsWith("/pkg.apk")) {
                                 newFile = "/data/app/" + app.packageName + ".apk";
                                 mvcmd = "busybox mv " + app.sourceDir + " " + newFile;
                             } else {
-                                newFile = app.sourceDir.replace(MoverActivity.SYSTEM_APP_FOLDER, "/data/app/");
+                                newFile = app.sourceDir
+                                        .replace(MoverActivity.SYSTEM_APP_FOLDER, "/data/app/");
                                 mvcmd = "busybox mv " + app.sourceDir + " /data/app/";
                             }
                         }
-                        if (BuildConfig.DEBUG)
-                            Logger.log("command: " + mvcmd);
+                        if (BuildConfig.DEBUG) Logger.log("command: " + mvcmd);
 
                         output = RootTools.sendShell(mvcmd, 10000);
 
@@ -199,8 +193,7 @@ public class AppClickListener implements OnItemClickListener {
                                     error += "\n" + str;
                                 }
                             }
-                            if (BuildConfig.DEBUG)
-                                Logger.log(error);
+                            if (BuildConfig.DEBUG) Logger.log(error);
                             ap.activity.showErrorDialog(error);
                         } else {
 
@@ -216,7 +209,8 @@ public class AppClickListener implements OnItemClickListener {
                             if (f.length() > 1) {
 
                                 if (!alreadySys) {
-                                    output = RootTools.sendShell("busybox chmod 644 " + newFile, 5000);
+                                    output = RootTools
+                                            .sendShell("busybox chmod 644 " + newFile, 5000);
                                     if (BuildConfig.DEBUG) {
                                         for (String str : output) {
                                             Logger.log(str);
@@ -227,17 +221,23 @@ public class AppClickListener implements OnItemClickListener {
                                 view.setVisibility(View.GONE);
                                 view.setTag("MOVED");
                                 AlertDialog.Builder b2 = new AlertDialog.Builder(ap.activity);
-                                b2.setMessage(appName + " successfully moved, you need to reboot your device.\nReboot now?");
-                                if (BuildConfig.DEBUG)
-                                    Logger.log("successfully moved");
+                                b2.setMessage(appName +
+                                        " successfully moved, you need to reboot your device.\nReboot now?");
+                                if (BuildConfig.DEBUG) Logger.log("successfully moved");
                                 b2.setPositiveButton(android.R.string.yes, new OnClickListener() {
                                     @Override
                                     public void onClick(final DialogInterface dialog, int which) {
-                                        if (BuildConfig.DEBUG)
-                                            Logger.log("reboot now");
-                                        ap.activity.sendBroadcast(new Intent("de.j4velin.ACTION_SHUTDOWN"));
+                                        if (BuildConfig.DEBUG) Logger.log("reboot now");
+                                        ap.activity.sendBroadcast(
+                                                new Intent("de.j4velin.ACTION_SHUTDOWN"));
                                         try {
-                                            RootTools.sendShell("am broadcast -a android.intent.action.ACTION_SHUTDOWN", 5000);
+                                            dialog.dismiss();
+                                        } catch (Exception e) {
+                                        }
+                                        try {
+                                            RootTools.sendShell(
+                                                    "am broadcast -a android.intent.action.ACTION_SHUTDOWN",
+                                                    5000);
                                             try {
                                                 Thread.sleep(1000);
                                             } catch (InterruptedException e) {
@@ -250,8 +250,7 @@ public class AppClickListener implements OnItemClickListener {
                                 b2.setNegativeButton(android.R.string.no, new OnClickListener() {
                                     @Override
                                     public void onClick(final DialogInterface dialog, int which) {
-                                        if (BuildConfig.DEBUG)
-                                            Logger.log("no reboot");
+                                        if (BuildConfig.DEBUG) Logger.log("no reboot");
                                         try {
                                             dialog.dismiss();
                                         } catch (Exception e) {
@@ -266,15 +265,13 @@ public class AppClickListener implements OnItemClickListener {
                     } catch (Exception e) {
                         ap.activity.showErrorDialog(e.getClass().getName() + " " + e.getMessage());
                         e.printStackTrace();
-                        if (BuildConfig.DEBUG)
-                            Logger.log(e);
+                        if (BuildConfig.DEBUG) Logger.log(e);
                     } finally {
                         RootTools.remount("/system", "ro");
                         RootTools.remount("/mnt", "ro");
                     }
                 } else {
-                    if (BuildConfig.DEBUG)
-                        Logger.log("can not remount /system");
+                    if (BuildConfig.DEBUG) Logger.log("can not remount /system");
                     ap.activity.showErrorDialog("Could not remount /system");
                 }
             }
